project/requirements.txt
celery==5.3.6
certifi==2023.11.17
coreschema==0.0.4
Django==5.0.1
django-cors-headers==4.3.1
django-environ==0.9.0
django-filter==23.5
django-timezone-field==6.1.0
djangorestframework==3.14.0
drf-yasg==1.20.0
environ==1.0
gunicorn==20.1.0
Markdown==3.4.4
packaging==23.2
pillow==10.2.0
requests==2.31.0
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4

